test(rivetkit-napi): move private tests to crate root
