+++ rivetkit-typescript/packages/rivetkit-napi/Cargo.toml
@@ -4,6 +4,7 @@ version.workspace = true
 edition.workspace = true
 authors.workspace = true
 license.workspace = true
+autotests = false
 
 [lib]
 crate-type = ["cdylib"]
