+++ rivetkit-typescript/packages/rivetkit-napi/src/actor_context.rs
@@ -911,23 +911,7 @@ fn js_http_request_to_core_request(request: JsHttpRequest) -> napi::Result<CoreR
 	.map_err(napi_anyhow_error)
 }
 
+// Test shim keeps moved tests in crate-root tests/ with private-module access.
 #[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn reset_runtime_state_clears_end_reason_without_touching_core_lifecycle_flags() {
-		let shared = ActorContextShared::default();
-		let ctx = CoreActorContext::new("actor-test", "actor", Vec::new(), "local");
-
-		ctx.set_started(true);
-		shared.set_end_reason(EndReason::Sleep);
-		assert!(shared.has_end_reason());
-		assert!(ctx.started());
-
-		shared.reset_runtime_state();
-
-		assert!(!shared.has_end_reason());
-		assert!(ctx.started());
-	}
-}
+#[path = "../tests/actor_context.rs"]
+mod tests;
